Separated DLQ reader tests that regards deletion of completed segments in its own suite
